% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt.R
\name{bbt_cite}
\alias{bbt_cite}
\alias{bbt_bib}
\alias{bbt_cayw}
\alias{bbt_bib_zotero}
\title{Interactively insert references from Zotero}
\usage{
bbt_cite(format = c("pandoc", "latex", "cite"), .action = bbt_print)

bbt_bib(translator = c("biblatex", "bibtex", "csljson"),
  .action = bbt_print)

bbt_cayw(..., .action = bbt_print)

bbt_bib_zotero(translator = c("biblatex", "bibtex", "csljson"),
  .action = bbt_print)
}
\arguments{
\item{format}{The citation format to use (in-text citations)}

\item{.action}{Use \link{bbt_return} to return the value without printing.}

\item{translator}{The translator to use (bibliography files)}

\item{...}{Pass anything to the cite-as-you-write endpoint (see link above)}
}
\value{
A character vector of length 1, the result of \code{.action}.
}
\description{
Using BetterBibTex's cite-as-you-write tool (
\url{https://retorque.re/zotero-better-bibtex/cayw/}). Citation text
is printed to the console by default (use the addin to insert directly
into the editor!). Use \code{bbt_bib_zotero} to use the current
selection in Zotero to generate a bibliography.
}
\examples{
if(interactive()) bbt_cite()
if(interactive()) bbt_bib()
if(interactive()) bbt_cayw(format = "mmd")
if(interactive()) bbt_bib_zotero()

}
