% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt.R
\name{bbt_insert}
\alias{bbt_insert}
\alias{bbt_print}
\alias{bbt_return}
\alias{bbt_copy}
\title{Do something with a value from the API}
\usage{
bbt_insert(text)

bbt_print(text)

bbt_return(text)

bbt_copy(text)
}
\arguments{
\item{text}{A character vector as lines}
}
\value{
the input
}
\description{
Note that insert uses the RStudio API, but will print values if the API is unavailable.
}
