% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt.R
\name{bbt_ref_cayw}
\alias{bbt_ref_cayw}
\alias{bbt_bib_cayw}
\title{Insert references from Zotero}
\usage{
bbt_ref_cayw(format = c("pandoc", "latex", "cite"), .action = bbt_print)

bbt_bib_cayw(
  translator = c("biblatex", "bibtex", "csljson", "cslyaml"),
  .action = bbt_print
)
}
\arguments{
\item{format}{The citation format to use (in-text citations)}

\item{.action}{Use \code{\link[=bbt_return]{bbt_return()}} to return the value without printing.}

\item{translator}{The translator to use (bibliography files)}
}
\value{
A character vector of length 1, the result of \code{.action}.
}
\description{
Using BetterBibTex's
\href{https://retorque.re/zotero-better-bibtex/cayw/}{cite-as-you-write tool}.
Unfortunately, this doesn't work well on some platforms.
Citation text
is printed to the console by default (use the addin to insert directly
into the editor!).
}
