% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{bbt_update_bib}
\alias{bbt_update_bib}
\alias{bbt_guess_bib_file}
\title{Update bibliography file}
\usage{
bbt_update_bib(
  path_rmd,
  path_bib = NULL,
  ignore = character(0),
  translator = NULL,
  library_id = getOption("rbbt.default.library_id", 1),
  overwrite = TRUE,
  filter = identity,
  quiet = FALSE
)

bbt_guess_bib_file(path_rmd, encoding = "UTF-8")
}
\arguments{
\item{path_rmd}{The path to the RMarkdown file(s).}

\item{path_bib}{Optionally, the path to the bibliography file.
The default value \code{NULL}uses \code{bbt_guess_bib_file()} to detect the path from
the 'YAML' front matter of \code{path_rmd}.}

\item{ignore}{A character vector of keys to disregard (useful if
\code{\link[=bbt_detect_citations]{bbt_detect_citations()}} gives spurious output).}

\item{translator}{Type of bibliography file to create. Options are \code{json}
(CSL-JSON), \code{biblatex} (BibLaTeX), \code{bibtex} (BibTeX), and \code{yaml}
(CSL YAML). CSL-JSON is recommended if users are not specifically using
a LaTeX citation processor.}

\item{library_id}{You may have to pass a specific library ID
if your options are not set to use globally unique keys.
Set the rbbt.default.library_id to ensure this value is
used by the addin. You can use \code{\link[=bbt_library_id]{bbt_library_id()}} to look
up this value by name.}

\item{overwrite}{Use \code{TRUE} to overwrite an existing file at \code{path}}

\item{filter}{Pass a function that can be used to modify the output
of \code{\link[=bbt_bib]{bbt_bib()}} before it is written to \code{path}.}

\item{quiet}{Use \code{TRUE} to suppress message on successful write.}

\item{encoding}{Passed to \code{\link[rmarkdown:yaml_front_matter]{rmarkdown::yaml_front_matter()}}}
}
\value{
\code{path_bib}, invisibly
}
\description{
This function takes an RMarkdown file location and a bibliography file
as inputs and updates the bibliography file.
}
