% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-bib.R
\name{bbt_write_bib}
\alias{bbt_write_bib}
\title{Write a bibliography file from Zotero citation keys}
\usage{
bbt_write_bib(
  path,
  keys = bbt_detect_citations(),
  ignore = character(),
  translator = bbt_guess_translator(path),
  library_id = getOption("rbbt.default.library_id", 1),
  overwrite = FALSE,
  filter = identity
)
}
\arguments{
\item{path}{The output path for the bibliography file. Must end with \code{.json}, \code{.bib}, or \code{.yaml}.}

\item{keys}{A character vector of citation keys.}

\item{ignore}{A character vector of keys to disregard (useful if
\code{\link[=bbt_detect_citations]{bbt_detect_citations()}} gives spurious output).}

\item{translator}{Type of bibliography file to create. Options are \code{json}
(CSL-JSON), \code{biblatex} (BibLaTeX), \code{bibtex} (BibTeX), and \code{yaml}
(CSL YAML). CSL-JSON is recommended if users are not specifically using
a LaTeX citation processor.}

\item{library_id}{You may have to pass a specific library ID
if your options are not set to use globally unique keys.
Set the rbbt.default.library_id to ensure this value is
used by the addin. You can use \code{\link[=bbt_library_id]{bbt_library_id()}} to look
up this value by name.}

\item{overwrite}{Use \code{TRUE} to overwrite an existing file at \code{path}}

\item{filter}{Pass a function that can be used to modify the output
of \code{\link[=bbt_bib]{bbt_bib()}} before it is written to \code{path}.}
}
\value{
The value of \code{path} (so that this function
can be used in the 'bibliography' RMarkdown YAML field)
}
\description{
Write a bibliography file to disk from a set of citation keys.
To automatically
generate a bibliography file from the items cited in your document when you
Knit, place a line like this one in your RMarkdown YAML header:

\code{bibliography: "`r rbbt::bbt_write_bib('bibliography.json', overwrite = TRUE)`"}
}
\section{Note}{
 Most users should use CSL-JSON format (\code{.json}) for their bibliographies.

RMarkdown's citation formatting uses
\href{https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_styles}{CSL styles}.
These styles require CSL-JSON data. RMarkdown converts other data formats
(e.g., BibLaTeX) to CSL-JSON. This is not lossless, and references other than
journal articles may be inaccurate. This is the case even if outputting to
PDF or TeX format.

Only use BibLaTeX or BibTeX if you are using
\href{https://pandoc.org/MANUAL.html#citation-rendering}{pandoc arguments to specify an alternative citation engine for raw TeX output}.

The default translator can be changed. For example:
\code{options(rbbt.default.translator = "json")}
}

\examples{
\dontrun{
bbt_write_bib("bibliography.json") # For CSL-JSON output (recommended)
bbt_write_bib("bibliography.bib")  # For BibLaTeX output
}

}
