% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{has_bbt}
\alias{has_bbt}
\alias{assert_bbt}
\title{Test for the Better BibTeX HTTP Endpoint}
\usage{
has_bbt()

assert_bbt()
}
\value{
\code{TRUE} if the endpoint exists, \code{FALSE} otherwise
}
\description{
Test for the Better BibTeX HTTP Endpoint
}
\examples{
has_bbt()

}
