% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt.R
\name{bbt_bib_selected}
\alias{bbt_bib_selected}
\alias{bbt_bib}
\title{Fetch bibliographical information from Zotero}
\usage{
bbt_bib_selected(
  translator = getOption("rbbt.default.translator", "biblatex"),
  .action = bbt_print
)

bbt_bib(
  keys,
  translator = getOption("rbbt.default.translator", "biblatex"),
  library_id = getOption("rbbt.default.library_id", 1),
  .action = bbt_print
)
}
\arguments{
\item{translator}{Type of bibliography file to create. Options are \code{json}
(CSL-JSON), \code{biblatex} (BibLaTeX), \code{bibtex} (BibTeX), and \code{cslyaml}
(CSL YAML). CSL-JSON is recommended if users are not specifically using
a LaTeX citation processor.}

\item{.action}{Use \code{\link[=bbt_return]{bbt_return()}} to return the value without printing.}

\item{keys}{A character vector of citation keys.}

\item{library_id}{You may have to pass a specific library ID
if your options are not set to use globally unique keys.
Set the rbbt.default.library_id to ensure this value is
used by the addin. You can use \code{\link[=bbt_library_id]{bbt_library_id()}} to look
up this value by name.}
}
\value{
The result of \code{.action}.
}
\description{
Fetch bibliographical information from your Zotero library. \code{bbt_bib()} fetches
information for a vector of citation keys; \code{bb_bib_selected()} fetches information
for the currently selected items in the Zotero pane.
}
\section{Note}{
 Most users should use CSL-JSON format (\code{.json}) for their bibliographies.

RMarkdown's citation formatting uses
\href{https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_styles}{CSL styles}.
These styles require CSL-JSON data. RMarkdown converts other data formats
(e.g., BibLaTeX) to CSL-JSON. This is not lossless, and references other than
journal articles may be inaccurate. This is the case even if outputting to
PDF or TeX format.

Only use BibLaTeX or BibTeX if you are using
\href{https://pandoc.org/MANUAL.html#citation-rendering}{pandoc arguments to specify an alternative citation engine for raw TeX output}.

The default translator can be changed. For example:
\code{options(rbbt.default.translator = "json")}
}

