% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt.R
\name{bbt_bib_selected}
\alias{bbt_bib_selected}
\alias{bbt_bib}
\title{Insert selected bibliography items from Zotero}
\usage{
bbt_bib_selected(
  translator = c("biblatex", "bibtex", "csljson", "cslyaml"),
  .action = bbt_print
)

bbt_bib(
  keys,
  translator = c("biblatex", "bibtex", "csljson", "cslyaml"),
  .action = bbt_print
)
}
\arguments{
\item{translator}{The translator to use (bibliography files)}

\item{.action}{Use \code{\link[=bbt_return]{bbt_return()}} to return the value without printing.}
}
\value{
The result of \code{.action}.
}
\description{
Returns the bibliography version of the selected items in the
Zotero pane.
}
