% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-citations.R
\name{bbt_detect_citations}
\alias{bbt_detect_citations}
\alias{bbt_guess_citation_context}
\title{Detect pandoc-style citations}
\usage{
bbt_detect_citations(
  path = bbt_guess_citation_context(),
  locale = readr::default_locale()
)

bbt_guess_citation_context()
}
\arguments{
\item{path}{A character vector, file or URL whose contents may contain citation keys.}

\item{locale}{See \code{\link[readr:default_locale]{readr::default_locale()}}. Use if encoding might
be a problem.}
}
\value{
A character vector of unique citation keys.
}
\description{
Detect pandoc-style citations
}
\examples{
bbt_detect_citations("\n@citation1 and [@citation2] but not \\\\@citation3")

}
