% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt-libraries.R
\name{bbt_libraries}
\alias{bbt_libraries}
\title{Query Zotero Libraries}
\usage{
bbt_libraries()
}
\value{
A data.frame with columns \code{id} and \code{name}.
}
\description{
Useful for specifying the \code{library_id} in \code{\link[=bbt_bib]{bbt_bib()}} or \code{\link[=bbt_write_bib]{bbt_write_bib()}}
when using a group library whose \code{id} may be different between users.
}
\examples{
if (has_bbt()) bbt_libraries()

}
