% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{bbt_update_bib}
\alias{bbt_update_bib}
\title{Update bibliography file}
\usage{
bbt_update_bib(path_rmd, path_bib, ...)
}
\arguments{
\item{path_rmd}{The path to the RMarkdown file.}

\item{path_bib}{The path to the bibliography file.}
}
\value{
The location of the bibliography file that has been updated
}
\description{
This function takes an RMarkdown file location and a bibliography file
as inputs and updates the bibliography file.
}
\details{
It also takes named arguments that can be passed to \code{\link[=bbt_write_bib]{bbt_write_bib()}}.
}
