% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-bib.R
\name{bbt_write_bib}
\alias{bbt_write_bib}
\title{Write a bibliography file from citation keys}
\usage{
bbt_write_bib(
  path,
  keys = bbt_detect_citations(),
  ignore = character(),
  translator = c("biblatex", "bibtex", "csljson", "cslyaml"),
  overwrite = FALSE
)
}
\arguments{
\item{path}{A path where the bibliography should be written}

\item{keys}{A character vector of citation keys.}

\item{ignore}{A character vector of keys to disregard (useful if
\code{\link[=bbt_detect_citations]{bbt_detect_citations()}} gives spurious output).}

\item{translator}{The translator to use (bibliography files)}

\item{overwrite}{Use \code{TRUE} to overwrite an existing file at \code{path}}
}
\value{
\code{path} (so that this can be used in the 'biblio' RMarkdown YAML field)
}
\description{
Write a bibliography file from citation keys
}
